requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==21.4.0
chardet==3.0.4
discord.py==1.7.3
multidict==6.0.2
typing-extensions==4.0.1
yarl==1.7.2
pymongo==4.0.1
dnspython==2.2.0